pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iras-scraper-v1"
version = "1.0.0"
description = "High-performance IRAS web scraper with advanced ReCAPTCHA v2 solving and batch processing capabilities"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
keywords = ["web-scraping", "selenium", "recaptcha", "iras"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Internet :: WWW/HTTP :: Browsers",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
  "selenium>=4.8.0,<5.0.0",
  "selenium-recaptcha-solver>=1.8.0",
  "openpyxl>=3.1.0",
  "pandas>=1.5.0",
  "requests>=2.28.0",
  "python-dotenv>=1.0.0",
  "webdriver-manager>=4.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.5", "black>=22.0.0", "flake8>=5.0.0", "mypy>=1.0.0"]

[project.urls]
Homepage = "https://github.com/jichong-tay/iras-scraper-v1"
Repository = "https://github.com/jichong-tay/iras-scraper-v1.git"
Issues = "https://github.com/jichong-tay/iras-scraper-v1/issues"

[project.scripts]
iras-scraper = "iras_scraper.main:main"

[tool.hatch.build.targets.wheel]
packages = ["iras_scraper"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short --strict-markers"
filterwarnings = [
  "ignore::RuntimeWarning:pydub.utils",
  "ignore::DeprecationWarning",
]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
  "unit: marks tests as unit tests",
  "webdriver: marks tests that require browser/WebDriver",
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.hatch.version]
path = "iras_scraper/__init__.py"
